requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
gunicorn==22.0.0
PyYAML==6.0.2
pytest==8.3.2
requests>=2.32.0
python-dotenv>=1.0.1
pydantic==2.8.2
httpx==0.27.0
composio==0.8.14
python-multipart==0.0.9
pdfplumber>=0.10.0
pandas>=2.2.0
openpyxl>=3.1.2
PyPDF2>=3.0.0
tqdm>=4.66.0
rich>=13.7.0
asyncpg>=0.29.0